Generate HTML-based Volcano Plot

Generate HTML-based Volcano Plot

Usage

plotVolcanoHTML(data, FC, p.vec, cutoff, fc.cutoff, main, x.lab, ...)

# S3 method for data.frame
plotVolcanoHTML(
  data,
  FC = signed.log2.fold.change,
  p.vec = p.value,
  cutoff = 0.05/nrow(data),
  fc.cutoff = 1,
  main = NULL,
  x.lab = NULL,
  target.labels,
  ...
)

# S3 method for stat_table
plotVolcanoHTML(
  data,
  FC = signed.log2.fold.change,
  p.vec = p.value,
  cutoff = 0.05/nrow(data$stat.table),
  fc.cutoff = 1,
  main = NULL,
  x.lab = NULL,
  apt.data,
  ...
)

Arguments

data

One of two object classes:

data.frame:

Containing log2-transformed fold-changes and corresponding p-values. In this case, an optional third column containing Target Names can be passed if specified in the target.labels = argument.

stat_table:

An object of class c(stat_table, *_table). When this argument is provided, the apt.data argument must also be passed.

FC

An un-quoted variable name in data for the fold-changes.

p.vec

An un-quoted variable name in data for the p-values.

cutoff

Horizontal cutoff for coloring points. Defaults to Bonferroni corrected level of significance.

fc.cutoff

Vertical cutoff for coloring coloring points (i.e. the fold-change). Defaults to doubling in fold-change.

main

Character. Main title for the plot. See ggtitle() for ggplot2 style graphics.

x.lab

Character. Optional string for the x-axis. Otherwise one is automatically generated (default).

...

Additional arguments passed to plotly::plotly().

target.labels

Optional for the data.frame method. An un-quoted variable name in data for the Target Names.

apt.data

An apt.data object containing target information used to generate point labels in the final plot. This object can be obtained from a call to getAnalyteInfo() on the original data.

Methods (by class)

plotVolcanoHTML(data.frame): Plot method for objects of class data.frame.

plotVolcanoHTML(stat_table): Plot method for objects of class stat_table.

See also

Author

Leigh Alexander, Stu Field

Examples

# Dummy up a fake table with minimal variables
data <- SomaDataIO::example_data
apts <- SomaDataIO::getAnalytes(data)
df <- withr::with_seed(100,
   data.frame(
              AptName = sample(apts, 50),
              t.stat = runif(50, 10, 20),
              p.value = runif(50),
              signed.log2.fold.change = rnorm(50),
              Zscore = rnorm(50),
              # generate fake target names
              target = paste0("Target", "_", sample(LETTERS, 50, replace = TRUE),
                              sample(1:50, 50), "-",
                              sample(letters, 50, replace = TRUE))
             )
)

# S3 `data.frame` method
# No Target labels -> `NA`
plotVolcanoHTML(df, cutoff = 0.1, fc.cutoff = 0.5)

# Add Target labels -> `targets`
plotVolcanoHTML(df, cutoff = 0.1, fc.cutoff = 0.5, target.labels = target)

# S3 `stat_table` method; calc.* family
# Dummy up a fake `stat_table` object
stbl <- list()
stbl$stat.table <- SomaDataIO::col2rn(df, "AptName")
stbl$test <- "Fake t-test"
stbl$data.dim <- dim(df)
stbl$y.response <- "LHS"
stbl$counts <- c(GroupA = 25, GroupB = 25)
stbl$log <- TRUE
stbl$data.frame <- "df"
class(stbl) <- c("stat_table", class(stbl))

# Pick up apt.data
ad <- SomaDataIO::getAnalyteInfo(data)

# Target names picked up from `apt.data`
plotVolcanoHTML(stbl, apt.data = ad, cutoff = 0.1, fc.cutoff = 0.5)

# Change FC arg on fly -> `Zscore`
# Pass args via '...' -> `x.lab`
plotVolcanoHTML(stbl, apt.data = ad, cutoff = 0.1, fc.cutoff = 0.5,
                FC = Zscore, x.lab = "Slope")