SomaLogic Primary Color Palette

This is the primary color palette used in SomaPlotr. It is used in the various scale_*_soma() functions and forms the basis of the SomaLogic theme.

Usage

palette_soma(n, random_order = FALSE)

scale_color_soma(...)

scale_fill_soma(...)

scale_continuous_color_soma(
  low = soma_colors2$pink,
  high = soma_colors2$blue,
  ...
)

scale_continuous_fill_soma(
  low = soma_colors2$pink,
  high = soma_colors2$blue,
  ...
)

Source

Influenced by: https://github.com/jrnold/ggthemes

Arguments

n

number of colors. This palette is set to the soma_colors2 object, which is currently 8 elements in length. The palette is recycled as necessary.

random_order

Logical. If the palette should be randomly sampled prior to rendering each plot to generate more colorful plots.

...

Arguments passed to ggplot2::discrete_scale(), ggplot2::scale_fill_gradient(), or ggplot2::scale_color_gradient().

low, high

Colours for low and high ends of the gradient.

Functions

scale_color_soma(): color discrete scale for SomaLogic graphics.

scale_fill_soma(): fill discrete scale for SomaLogic graphics.

scale_continuous_color_soma(): color continuous scale for SomaLogic graphics.

scale_continuous_fill_soma(): fill continuous scale for SomaLogic graphics.

See also